+++ inscripta/biocantor/io/parser.py
@@ -83,7 +83,7 @@ def seq_to_parent(
     """Convert a string into a Parent object. This is the intermediate that transfers a BioPython sequence object to
     a BioCantor sequence object.
 
-    NOTE: This sequnece is assumed to be the entire chromosome.
+    NOTE: This sequence is assumed to be the entire chromosome.
 
     Args:
         seq: String of sequence.
@@ -122,11 +122,13 @@ def seq_chunk_to_parent(
     Returns:
         An instantiated Parent object ready to be passed to a constructor.
     """
+    chunk_id = f"{sequence_name}:{start}-{end}"
     return Parent(
-        id=f"{sequence_name}:{start}-{end}",
+        id=chunk_id,
         sequence=Sequence(
             seq,
             alphabet,
+            id=chunk_id,
             type=SequenceType.SEQUENCE_CHUNK,
             parent=Parent(
                 location=SingleInterval(